msm: idle-v7: Add no-ops for cortex targets
For Cortex targets a delay of about 300 NOPs is required to prevent
a race condition that might cause the SPM state machine and the CPU to
be out of sync during failed power collapses. The NOPs were present
in code and conditional to a specific target type, make these common
to all Cortex targets.

Change-Id: Ibee691e168df412376f0a042d54a643c731ae0e3
Signed-off-by: Girish Mahadevan <girishm@codeaurora.org>