ARM: dts: msm: shut off usb clocks on suspend for 9625
So far, when the usb bus was in suspend state the usb driver used
wake up interrupts from the usb bam to in order to start remote wake up
signal to the host due to a packet arrival in the bam pipe. In order
to this interrupt to work, we had to keep usb clocks enable during
suspend.

Today, after implementing handshake with the IPA resource manager, upon
start of remote wake up, the consumer will be first requested by the
resource manager to signal the usb to exit from low power mode, before
the first packet is pushed into the bam pipe, so before the wake up
interrupt will be received.

This allow us to shutdown the usb clocks during suspend and therefore to
support XO shutdown during suspend.

Change-Id: Idd0acfa0a760ada10c25b7ef0be498724acff868
Signed-off-by: Ido Shayevitz <idos@codeaurora.org>