regulator: qpnp-regulator: Add support for boost-bypass type regulators
Add support for QPNP PMIC boost-bypass type regulators.  This
type of regulator is designed to output a configurable minimum
voltage.  If the input voltage is above the set point voltage,
then the regulator automatically enters bypass mode such that the
output voltage is equal to the input voltage.  If the input
voltage is below the set point voltage, then it is boosted up to
the set point voltage.

Change-Id: I93704339b7cb730fae099471bf8ec872dae54c04
Signed-off-by: David Collins <collinsd@codeaurora.org>