ARM:dts:msm: Remove CPU notifications to RPM for MSM8226v2
RPM uses SPM signals from L2 and each of the APPS core to
determine if it is okay to transition from active set to
sleep set. On 8x10 and 8x26 due to a Cortex-A7 ARM errata
where primary core(C0) when waking up from power collapse
has to wake up all the secondary cores to invalidate
all it's L1 duplicate tag RAMs and power collapse them
again. Due to this there is an unnecessary increase in
SPM-RPM handshake interrupts for secondary cores.

This patch removes the APPS cores notification to RPM in
RPM assisted PC and only L2 handshakes with RPM. This helps
to reduce the interrupt traffic generated by APPS cores to
RPM.

CRs-Fixed: 524102
Change-Id: Ieafacdab811d2571965b91e5772663e84a89b896
Signed-off-by: Murali Nalajala <mnalajal@codeaurora.org>