ARM: dts: msm: correct the PFM threshold for 8974 and 8974pro
The PFM threshold is set to 73 right now. This value although is
correct for 8974, needs to change to 76 for 8974pro. Changing the
value to 76 for 8974pro does not impact 8974 because 76 is the upper
threshold for 300Mhz frequency for all the bins of 8974pro
and 8974 and the lowest value for the next frequency for
all the bins of 8974 and 8974pro is 83. This guarantees that
PFM will be activated only in single core running at 300Mhz for
all bins of 8974 and 8974pro.

Change-Id: I9a7be1ee3c2affc43fd2538924d8827c9eb68ce2
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>