USB: HSIC SMSC HUB: Add support for 3502 model
The SMSC 3502 HUB has only one USB UICC port. "VBAT only" single
supply configuration is used. A gpio is used to control the
external oscillator which supplies the ref clock to the HUB. USB
UICC card is connected to this HUB on MSM8926. Provide a sysfs
interface for userspace to enable/disable the HUB. This is
required to save the power when a non-supported card is inserted.

The HUB is disabled during boot. If the card is inserted and supports
UICC, HUB is enabled by the userspace.

Change-Id: Ic2b0c5bd2066a2ecae7f18aac2ab8d0bc2f075c0
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>