crypto: msm: Fix key length setting
Hardcoding key length to a constant value results in overwriting
command element information in the command descriptor beyond what
is allocate for the key information in the command element list.
This results in corruption of the command list in the case where
command descriptors are used for configuring crypto registers and
needlessly writing to key registers that are not relevant to the
operation (when writing to registers directly), and thus leads to
crypto operation failures. Fix is to use the key length based on
what is requested by the client.

Change-Id: I2cdac15e2dbc849f14c59a3383e6e1098cae3a60
Signed-off-by: Zhen Kong <zkong@codeaurora.org>