msm: cpr-regulator: change sequence for handling CPR voltage over ceiling
The current CPR driver increases the UP threshold to its maximum value when
the ceiling voltage is hit, but it is not disabling the UP interrupt
because the CPR controller needs a NACK command from software if the new UP
threshold is crossed for any reason. Based on analysis of CPR hardware
behavior, disabling CPR UP interrupt when CPR hits the ceiling voltage
is not harmful even though the CPR controller waits for a response from
software indefinitely. The CPR controller restarts sampling again as soon
as software switches to new voltage corner if it had been halted due to
hitting the ceiling voltage.

Modify the cpr-regulator driver so that the CPR UP interrupt is disabled
for a given voltage corner when the ceiling voltage for that corner is
reached. Re-enable the UP interrupt for a corner if a DOWN interrupt is
ever received for the same corner.

CRs-fixed: 602095
Change-Id: I0f2c479464cc43f784e97422095df395fa0c7b47
Signed-off-by: Ke Liu <keliu@codeaurora.org>