power: qpnp-bms: use better defaults for power supply props
Currently the BMS driver just assumes that the battery power supply will
export whatever properties it requires. However, if an external charger
is being used, having good default values for the power supply
properties that are not exported by the external charger will improve
compatibility.

Change the status property to default to unknown, the present property
to default to true, and the online property to default to true if they
are not exported by the battery power supply.

Change-Id: Ic10429881f221cfb427720bdd73d908e87b912b0
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>