msm: pm: Acquire scm_handoff_lock conditionally
Acquisition of the scm_handoff_lock requires a TrustZone software
image that will release the lock. Without this, a deadlock will
occur since the lock will remain held forever.

Since this lock is only really needed if the qcom,allow-synced-levels
property is specified for the lpm-levels device, conditionally acquire
the lock based on this property. This avoid the TrustZone software
dependency when qcom,allow-synced-levels is not specified.

Conflicts:

	arch/arm/boot/dts/msm8610-v1-pm.dtsi
	arch/arm/boot/dts/msm8610-v2-pm.dtsi

Change-Id: If29907497d1edd89c9c7d97024784505bb5d5b5f
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>
Signed-off-by: Sivasri Kumar Vanka <sivasri@codeaurora.org>