power: qpnp-charger: override DCIN_LLIMIT when Parallel charge is disabled
In the parallel charge configuration there is a possibility
of DC_IN being detected as valid and thus allowing current
to flow on the DC_IN path when this is undesirable such
as USB being suspended.

Fix this by keeping the override in place when parallel
path charging is toggled to disable.

CRs-Fixed: 624440
Change-Id: I65d85584ddc800e6d24423707188adedf20681f4
Signed-off-by: David Keitel <dkeitel@codeaurora.org>