msm: vidc: recover from firmware fatal error
If firmware goes in to bad state, the video driver
should try to restart the firmware, so that subsequent
new video sessions can play fine without error.
Previous to this change, the video driver would
unload the firmware only after all open instances
have been closed by the client. If the client fails
to close any of the open instances, then firmware
would not be unloaded and continue to remain in bad
state. This would fail all new video sessions and could
be overcome only after the device was reset.

CRs-Fixed: 637679
Change-Id: Ia0e7e09fe64948b0ca420ba707381639d7d9a87d
Signed-off-by: Arun Menon <avmenon@codeaurora.org>