power: qpnp-charger: if OTG is on and DC_IN removed resume charging
In the case DC_IN is charging and an OTG device is connected
subsequently and the DC_IN charger removed, battery would be still
reported as full despite the fact that no charger is connected.

Fix this by checking if OTG is enabled in
qpnp_chg_dc_dcin_valid_irq_handler.

Change-Id: Id0f7588fdcfb86261e129ab5b73e7da587889713
CRs-Fixed: 654771
Signed-off-by: David Keitel <dkeitel@codeaurora.org>