usb: dwc3: gadget: Ignore L1 RESUME events
Events may get generated for Link Power Management L1
Suspend and Resume. The suspend case currently ignores
L1 and only proceeds to call the gadget driver suspend
routines if it is for L2/U3 entry. However, the resume
callback doesn't make this distinction, which will
result in the gadget layer resume getting called for
which there had not been a previous suspend.

CRs-Fixed: 739812
Change-Id: Ib38e8949ce78ab40bd4f78041de92103869081b6
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>