msm: camera: isp: Add support for AV time stamp for VT call
For vt call we need to update the camera buffers with AVTimer
timestamp instead of normal kernel time stamp to prevent
AV sync issues.
A new AVTimer driver API has been provided which directly provides
the timestamp value in ticks that can be converted to appropriate
timescale and tagged to the video frames.

Change-Id: I9923c066602819eed0ccd06e112f12cd85b3d71b
Signed-off-by: Suman Mukherjee <sumam@codeaurora.org>